# server/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: PubSubServer REST API
    description: REST API for PubSubServer
    version: "1.0"
servers:
    - url: http://localhost:8080
paths:
    /SendMessage/{roomName}/{message}:
        get:
            tags:
                - PubSubServer
            summary: sendMessage PubSubServer
            operationId: PubSubServer#sendMessage
            parameters:
                - name: roomName
                  in: path
                  description: roomName to publish
                  required: true
                  schema:
                    type: string
                    description: roomName to publish
                    example: room1
                  example: room1
                - name: message
                  in: path
                  description: message to publish
                  required: true
                  schema:
                    type: string
                    description: message to publish
                    example: hello
                  example: hello
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: string
                                example: Dolorem iure similique modi blanditiis consequatur totam.
                            example: Magni inventore.
    /publish/{roomName}:
        get:
            tags:
                - PubSubServer
            summary: publish PubSubServer
            operationId: PubSubServer#publish
            parameters:
                - name: roomName
                  in: path
                  description: roomName to publish
                  required: true
                  schema:
                    type: string
                    description: roomName to publish
                    example: room1
                  example: room1
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: string
                                example: Dolorem minima omnis eum quo ullam.
                            example: Ut et eum impedit iure.
    /subscribe/{roomName}:
        get:
            tags:
                - PubSubServer
            summary: subscribe PubSubServer
            operationId: PubSubServer#subscribe
            parameters:
                - name: roomName
                  in: path
                  description: roomName to subscribe
                  required: true
                  schema:
                    type: string
                    description: roomName to subscribe
                    example: room1
                  example: room1
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: string
                                example: Nobis non expedita laudantium odio.
                            example: Ut sit velit eligendi dolor.
components: {}
tags:
    - name: PubSubServer
      description: The PubSubServer service performs operations on messages.